requirements.txt
spacy==3.7.2
google-generativeai==0.3.0
python-dotenv==1.0.0
numpy>=1.24.0,<2.0.0
cymem>=2.0.2,<2.1.0
murmurhash>=1.0.2,<1.1.0
preshed>=3.0.2,<3.1.0
thinc>=8.1.11,<8.2.0
wasabi>=0.9.1
srsly>=2.4.3,<3.0.0
catalogue>=2.0.6,<3.0.0
typer>=0.9.0,<0.10.0
pydantic>=1.7.4,<3.0.0
jinja2>=2.11.3
requests>=2.28.0
smart-open>=5.2.1
langcodes>=3.2.0